Feeding Hungry animals 3D and scoring and saving high score
Player can move in directions and throw a food towards incoming animals to feed them. if player is not able to feed the animals more than three times then player health will become zero and game will be over. This game is developed using Unity and is 3D project, download and import the package to use it.